Improve display framebuffer usage now that we have more RAM
When we want to update only a rectangle of the screen, use a temporary
buffer. This costs 150KiB more, but we have that now.

The three benchmark display updates now take:

- 140 ms - full screen update
-  32 ms - small text move down
-  48 ms - text moved up a few lines

Not too shabby. And particularly the worst case -- updating only a
vertical line -- is now sped up by a lot.